Move preview links from a tab into a Share dialog in the site header
