+++ apps/ui/src/components/site-overview-view/style.module.css
@@ -15,29 +15,26 @@
 	color: var(--wpds-color-fg-content-neutral-weak);
 }
 
+/* The toolbar brings its own inline padding and pins its actions to the top of
+   whatever row hosts it, so this row only supplies the height — matching
+   .header in ui-classic/components/session-view/style.module.css keeps the
+   header identical either side of the chat/overview switch. */
 .header {
 	position: relative;
 	display: flex;
 	align-items: center;
+	height: 54px;
 	gap: var(--wpds-dimension-padding-sm);
-	padding-block: var(--wpds-dimension-padding-lg) var(--wpds-dimension-padding-sm);
-	padding-inline: var(--wpds-dimension-padding-lg);
 	font-size: var(--wpds-typography-font-size-sm);
 	color: var(--wpds-color-fg-content-neutral);
 	-webkit-app-region: drag;
 }
 
 /* Keep in sync with .headerSidebarCollapsed in
-   ui-classic/components/session-view/style.module.css so the site menu sits
-   in the same spot when switching between chat and overview. */
+   ui-classic/components/session-view/style.module.css so the site sits in the
+   same spot when switching between chat and overview. */
 .headerSidebarCollapsed {
-	padding-top: calc(
-		var(--wpds-dimension-padding-xl) + var(--wpds-dimension-padding-2xl) +
-			var(--wpds-dimension-padding-lg)
-	);
-	/* With the sidebar hidden the panel goes full-bleed, putting this header
-	   at the window's left edge — give the site menu more breathing room. */
-	padding-inline-start: var(--wpds-dimension-padding-lg);
+	padding-inline-start: 84px;
 }
 
 .tabsFrame {
