+++ apps/ui/src/components/site-settings-view/index.tsx
@@ -32,7 +32,7 @@ import type { SupportedPHPVersion } from '@studio/common/types/php-versions';
 import type { DataFormControlProps, Field, Form } from '@wordpress/dataviews';
 import type { FormEvent } from 'react';
 
-type TabId = 'overview' | 'connections' | 'previews' | 'general' | 'debugging';
+type TabId = 'overview' | 'connections' | 'general' | 'debugging';
 
 interface FormData {
 	name: string;
@@ -323,7 +323,7 @@ export function SiteSettingsForm( { site, activeTab }: { site: SiteDetails; acti
 	);
 }
 
-const SITE_TABS: TabId[] = [ 'overview', 'connections', 'previews', 'general', 'debugging' ];
+const SITE_TABS: TabId[] = [ 'overview', 'connections', 'general', 'debugging' ];
 
 export function isSiteSettingsTab( value: string ): value is TabId {
 	return ( SITE_TABS as string[] ).includes( value );
