--- apps/ui/src/components/site-toolbar/connections-dialog.module.css
@@ -1,87 +1,0 @@
-.intro {
-	margin: 0 0 var(--wpds-dimension-padding-md);
-	font-size: var(--wpds-typography-font-size-sm);
-	color: var(--wpds-color-fg-content-neutral-weak);
-}
-
-.empty {
-	margin: 0;
-	padding: var(--wpds-dimension-padding-lg) 0;
-	font-size: var(--wpds-typography-font-size-sm);
-	color: var(--wpds-color-fg-content-neutral-weak);
-	text-align: center;
-}
-
-.list {
-	display: flex;
-	flex-direction: column;
-	gap: var(--wpds-dimension-gap-xs);
-	margin: 0;
-	padding: 0;
-	list-style: none;
-}
-
-.row {
-	display: flex;
-	align-items: center;
-	gap: var(--wpds-dimension-gap-sm);
-	padding: var(--wpds-dimension-padding-sm) var(--wpds-dimension-padding-md);
-	border: 1px solid var(--wpds-color-stroke-surface-neutral);
-	border-radius: var(--wpds-border-radius-md);
-}
-
-/* The connection the toolbar acts on carries the brand edge, so the answer to
-   "where does Push go" is visible without reading the rows. */
-.rowActive {
-	border-color: var(--wpds-color-stroke-interactive-brand);
-}
-
-.rowText {
-	display: flex;
-	flex-direction: column;
-	gap: 2px;
-	min-width: 0;
-	flex: 1;
-}
-
-.rowTitle {
-	display: flex;
-	align-items: center;
-	gap: var(--wpds-dimension-gap-xs);
-	font-size: var(--wpds-typography-font-size-sm);
-	font-weight: 600;
-	color: var(--wpds-color-fg-content-neutral);
-}
-
-.badge {
-	padding: 1px 6px;
-	border-radius: var(--wpds-border-radius-sm);
-	background: var(--wpds-color-bg-surface-info-weak);
-	color: var(--wpds-color-fg-content-info);
-	font-size: var(--wpds-typography-font-size-xs);
-	font-weight: var(--wpds-typography-font-weight-regular);
-}
-
-.rowUrl,
-.rowMeta {
-	font-size: var(--wpds-typography-font-size-xs);
-	color: var(--wpds-color-fg-content-neutral-weak);
-	white-space: nowrap;
-	overflow: hidden;
-	text-overflow: ellipsis;
-}
-
-.rowActions {
-	display: flex;
-	align-items: center;
-	gap: var(--wpds-dimension-gap-xs);
-	flex: 0 0 auto;
-}
-
-.disconnect {
-	color: var(--wpds-color-fg-content-error);
-}
-
-.addButton {
-	margin-block-start: var(--wpds-dimension-padding-md);
-}
