--- apps/ui/src/components/site-toolbar/preview-sites-dialog.module.css
@@ -1,59 +1,0 @@
-.intro {
-	margin: 0 0 var(--wpds-dimension-padding-md);
-	font-size: var(--wpds-typography-font-size-sm);
-	color: var(--wpds-color-fg-content-neutral-weak);
-}
-
-.empty {
-	margin: 0;
-	padding: var(--wpds-dimension-padding-lg) 0;
-	font-size: var(--wpds-typography-font-size-sm);
-	color: var(--wpds-color-fg-content-neutral-weak);
-	text-align: center;
-}
-
-.list {
-	display: flex;
-	flex-direction: column;
-	gap: var(--wpds-dimension-gap-xs);
-	margin: 0;
-	padding: 0;
-	list-style: none;
-}
-
-.row {
-	display: flex;
-	align-items: center;
-	gap: var(--wpds-dimension-gap-sm);
-	padding: var(--wpds-dimension-padding-sm) var(--wpds-dimension-padding-md);
-	border: 1px solid var(--wpds-color-stroke-surface-neutral);
-	border-radius: var(--wpds-border-radius-md);
-}
-
-.rowText {
-	display: flex;
-	flex-direction: column;
-	gap: 2px;
-	min-width: 0;
-	flex: 1;
-}
-
-.rowTitle {
-	font-size: var(--wpds-typography-font-size-sm);
-	color: var(--wpds-color-fg-content-neutral);
-	white-space: nowrap;
-	overflow: hidden;
-	text-overflow: ellipsis;
-}
-
-.rowMeta {
-	font-size: var(--wpds-typography-font-size-xs);
-	color: var(--wpds-color-fg-content-neutral-weak);
-}
-
-.rowActions {
-	display: flex;
-	align-items: center;
-	gap: var(--wpds-dimension-gap-xs);
-	flex: 0 0 auto;
-}
